Changed telling the user to click Quest 2/3 to "the base directory"
The user <@906398945964613643> has "Standalone HMD" (handheld mobile device) on their Quest 2 when looking at their files, instead of, well, Quest 2. Not sure why this happens, but to avoid that causing more confusion if the follower has something else, I've set it to be an umbrella term. If any more issues come from this, I'll add some clarification below it!